requirements.txt
Django
pandas
matplotlib
Pillow
openai
phyton-dotenv
numpy
requests